Removed the dependency from Diff to Parsing by moving ParserUtils to Utils
