+++ diff/pom.xml
@@ -20,11 +20,6 @@
         </dependency>
         <dependency>
             <groupId>com.tblf</groupId>
-            <artifactId>parsing</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.tblf</groupId>
             <artifactId>management</artifactId>
             <version>1.0.0-SNAPSHOT</version>
             <scope>test</scope>
