+++ utils/src/test/java/com/tblf/utils/ParserUtilsTest.java
@@ -1,4 +1,4 @@
-package com.tblf.parsing;
+package com.tblf.utils;
 
 import org.junit.Assert;
 import org.junit.Test;
